Moved KeystoreList to /components
